sip_connect/quantum_components.py
import math
import ephem
import hashlib
import requests
import socket
import time
from datetime import datetime
from qiskit import QuantumCircuit, transpile
from qiskit_aer import Aer
import numpy as np
import json

class MelodyQuantumGenerator:
    def __init__(self, melody_sequence):
        if melody_sequence is None:
            raise ValueError("melody_sequence cannot be None")
        self.melody = melody_sequence

    @staticmethod
    def calculate_quantum_modulation(note):
        base_angle = (note % 360) * np.pi / 180
        return math.cos(base_angle) * math.pi

    def generate_quantum_key(self):
        n_qubits = len(self.melody)
        qc = QuantumCircuit(n_qubits, n_qubits)

        for i, note in enumerate(self.melody):
            angle = self.calculate_quantum_modulation(note)
            qc.rx(angle, i)
            qc.h(i)

        for i in range(n_qubits - 1):
            qc.cx(i, i + 1)

        qc.measure(range(n_qubits), range(n_qubits))

        backend = Aer.get_backend('qasm_simulator')
        transpiled_circuit = transpile(qc, backend)
        result = backend.run(transpiled_circuit, shots=1).result().get_counts()

        quantum_key = list(result.keys())[0]
        return int(quantum_key, 2)


class EnvironmentalEntropy:
    def __init__(self):
        self.noaa_api_endpoint = "https://services.swpc.noaa.gov/json/rtsw/rtsw_mag_1m.json"
        self.network_nodes = ["8.8.8.8", "1.1.1.1", "208.67.222.222"]
        self.moon = ephem.Moon()
        self.jupiter = ephem.Jupiter()

    def get_astronomical_data(self):
        self.moon.compute(datetime.now())
        self.jupiter.compute(datetime.now())
        return {
            'lunar_phase': float(self.moon.phase),
            'jupiter_elong': float(self.jupiter.elong),
            'jupiter_phase': float(self.jupiter.phase),
            'jupiter_radius': float(self.jupiter.radius)
        }

    def measure_network_latency(self):
        latencies = {}
        for node in self.network_nodes:
            try:
                start = time.time()
                socket.create_connection((node, 53), timeout=1)
                latencies[node] = time.time() - start
            except socket.error:
                latencies[node] = 0
        return latencies

    def generate_entropy(self):
        try:
            response = requests.get(self.noaa_api_endpoint, timeout=5)
            response.raise_for_status()
            solar_data = response.json()[-1]
            solar = {
                'density': solar_data.get('density', 1),
                'speed': solar_data.get('speed', 1),
                'temperature': solar_data.get('temperature', 1)
            }
        except (requests.RequestException, json.JSONDecodeError, IndexError, KeyError):
            # Fallback values if API fails
            solar = {'density': 1, 'speed': 1, 'temperature': 1}

        astro = self.get_astronomical_data()
        latency = self.measure_network_latency()

        entropy_base = (
                solar['density'] *
                solar['speed'] *
                astro['jupiter_elong'] *
                astro['jupiter_phase'] *
                astro['jupiter_radius']
        )

        entropy_source = f"{entropy_base}{sum(latency.values())}"
        return int(hashlib.sha256(entropy_source.encode()).hexdigest(), 16)

# Constants
DEFAULT_MELODY = [440, 494, 523, 587, 659]

