sip_connect/requirements.txt
fastapi
uvicorn
websockets
qiskit
cryptography
pycryptodome
ephem
requests
# Other dependencies from your quantum implementation
numpy
pandas
python-multipart
paho-mqtt

